Preparation for OrderedRDD2.
Use off-heap storage.
Replaced Platform => is.hail.annotations.Memory.
Offsets are now Long, including stored offests to binary and arrays.

Add finalizer to MemoryBuffer.

Use codec to serialize UnsafeRow, UnsafeIndexedSeq.

Addressed comments.

Addressed comments 2.

Fixed test.

+++ src/main/scala/is/hail/utils/BytePacker.scala
@@ -3,13 +3,13 @@ package is.hail.utils
 import scala.collection.mutable
 
 class BytePacker {
-  val slots = new mutable.TreeSet[(Int, Int)]
+  val slots = new mutable.TreeSet[(Long, Long)]
 
-  def insertSpace(size: Int, start: Int) {
+  def insertSpace(size: Long, start: Long) {
     slots += size -> start
   }
 
-  def getSpace(size: Int, alignment: Int): Option[Int] = {
+  def getSpace(size: Long, alignment: Long): Option[Long] = {
 
     // disregard spaces smaller than size
     slots.iteratorFrom(size -> 0)
@@ -18,7 +18,7 @@ class BytePacker {
         val start = x._2
 
         // find a start position with the proper alignment
-        var i = 0
+        var i = 0L
         while (i + size <= spaceSize) {
           // space found
           if ((start + i) % alignment == 0) {
